+++ jwst/resample/resample.py
@@ -1,5 +1,4 @@
 import logging
-import re
 
 import numpy as np
 from drizzle import util
@@ -238,13 +237,6 @@ class ResampleData:
                 axis=0
             )
         )
-
-        # TODO: The following two methods and calls should be moved upstream to
-        # ResampleStep and ResampleSpecStep respectively
-        if isinstance(output_model, datamodels.ImageModel):
-            self.update_fits_wcs(output_model)
-        if isinstance(output_model, datamodels.SlitModel):
-            self.update_slit_metadata(output_model)
 
         self.update_exposure_times(output_model)
         self.output_models.append(output_model)
@@ -491,60 +483,3 @@ class ResampleData:
             wtscale=wtscale,
             fillstr=fillval
         )
-
-    def update_fits_wcs(self, model):
-        """
-        Update FITS WCS keywords of the resampled image.
-        """
-        # Delete any SIP-related keywords first
-        pattern = r"^(cd[12]_[12]|[ab]p?_\d_\d|[ab]p?_order)$"
-        regex = re.compile(pattern)
-
-        keys = list(model.meta.wcsinfo.instance.keys())
-        for key in keys:
-            if regex.match(key):
-                del model.meta.wcsinfo.instance[key]
-
-        # Write new PC-matrix-based WCS based on GWCS model
-        transform = model.meta.wcs.forward_transform
-        model.meta.wcsinfo.crpix1 = -transform[0].offset.value + 1
-        model.meta.wcsinfo.crpix2 = -transform[1].offset.value + 1
-        model.meta.wcsinfo.cdelt1 = transform[3].factor.value
-        model.meta.wcsinfo.cdelt2 = transform[4].factor.value
-        model.meta.wcsinfo.ra_ref = transform[6].lon.value
-        model.meta.wcsinfo.dec_ref = transform[6].lat.value
-        model.meta.wcsinfo.crval1 = model.meta.wcsinfo.ra_ref
-        model.meta.wcsinfo.crval2 = model.meta.wcsinfo.dec_ref
-        model.meta.wcsinfo.pc1_1 = transform[2].matrix.value[0][0]
-        model.meta.wcsinfo.pc1_2 = transform[2].matrix.value[0][1]
-        model.meta.wcsinfo.pc2_1 = transform[2].matrix.value[1][0]
-        model.meta.wcsinfo.pc2_2 = transform[2].matrix.value[1][1]
-        model.meta.wcsinfo.ctype1 = "RA---TAN"
-        model.meta.wcsinfo.ctype2 = "DEC--TAN"
-
-        # Remove no longer relevant WCS keywords
-        rm_keys = ['v2_ref', 'v3_ref', 'ra_ref', 'dec_ref', 'roll_ref',
-                   'v3yangle', 'vparity']
-        for key in rm_keys:
-            if key in model.meta.wcsinfo.instance:
-                del model.meta.wcsinfo.instance[key]
-
-    def update_slit_metadata(self, model):
-        """
-        Update slit attributes in the resampled slit image.
-
-        This is needed because model.slit attributes are not in model.meta, so
-        the normal update() method doesn't work with them. Updates output_model
-        in-place.
-        """
-        for attr in ['name', 'xstart', 'xsize', 'ystart', 'ysize',
-                     'slitlet_id', 'source_id', 'source_name', 'source_alias',
-                     'stellarity', 'source_type', 'source_xpos', 'source_ypos',
-                     'dispersion_direction', 'shutter_state']:
-            try:
-                val = getattr(self.input_models[-1], attr)
-            except AttributeError:
-                pass
-            else:
-                if val is not None:
-                    setattr(model, attr, val)